requirements.txt
asgiref==3.11.1
cpf_cnpj==0.1.4
Django==6.0.4
django-cpf-cnpj==1.0.0
django-localflavor==5.0
django-phonenumber-field==8.4.0
djangorestframework==3.17.1
phonenumbers==9.0.27
python-stdnum==2.2
sqlparse==0.5.5
tzdata==2026.1
